Improve mounting of data disks
The plugin now has a custom version of the script used by Altus Director
to mount unmounted volumes, in this case, data disks such as local
SSDs. The following two improvements are included.

* Disks whose device names end in a digit are now found and mounted,
  and not skipped. This is necessary for NVMe devices, which show up
  as /dev/nvme0n1, /dev/nvme0n2, and so on. This same improvement was
  already made to the AWS plugin some time ago.
* Entries for mounted data disks in /etc/fstab are now listed by
  block ID / UUID. In AWS, it was observed that the mount order for
  persistent disks could be changed on instance restart, so tracking
  by UUID avoids this problem. (For ephemeral storage, it doesn't
  matter.)

The mere presence of the script is sufficient for Altus Director to use
it instead of its internal default.