Use shorter hostnames for GCE instances
Previously, the name for a new GCE instance was computed as the
combination of a prefix in its instance template and its Director
instance ID, which is a UUID. This causes the name to be quite long,
often leading to an instance hostname longer than the maximum of 64
characters. Now, instead of using a UUID directly, the base 32 encoding
of the UUID bytes is used in the name. This saves 10 characters: 26
instead of 36. Base 64 cannot be used because Linux hostnames can only
use lowercase a - z, digits 0 - 9, and hyphens.