shared VPC support in the network validation
Shared VPC support was added in https://github.com/cloudera/director-google-plugin/pull/161. This allowed for the possibility that the network resource lives in a separate project from the project where we want to create the compute instances. We need to add logic to this check to utilize the new `NETWORK_PROJECT` variable that we introduced so that we don't look in the wrong project for the network.